requirements.txt
# Requirements for clackbot

wheel
python-dotenv
discord
emoji
PyNaCl
requests